Added rotating Objects, FpsCounter and updated Camera

+++ app/src/main/java/adrians/game/model/GameObject.java
@@ -6,6 +6,6 @@ import adrians.framework.util.Painter;
  * Created by pierre on 10/02/16.
  */
 public abstract class GameObject {
-    public abstract void update(double delta);
+    public abstract void update(float delta);
     public abstract void render(Painter g);
 }
